add lesson3.1
